상점 등록, 상점 조회, 리뷰 등록 추가

+++ src/main/java/com/app/grip/utils/jwt/JwtService.java
@@ -63,6 +63,6 @@ public class JwtService {
         }
 
         // 3. userId 추출
-        return claims.getBody().get("userNo", Long.class);
+        return (long) claims.getBody().get("userNo", Integer.class);
     }
 }
